Fix bot repeating staging status in alternating messages
When multiple failing tests on the staging webpage are found, the status
message string could alternate as an unordered dict was used causing the bot
to constantly repeating status messages when in defcon 5 it should be less
chatty. This is solved by always sorting the keys in the hash.

Also add tests to ensure this by mocking up web calls.

Fixes #10